avatar selection is now a component
